fix: kalau udah ada pages/index.md, abaikan membuat pages/index.html

+++ app.mjs
@@ -105,7 +105,7 @@ for (let lib of libraryCss) {
 }
 membuatFile("index.html", BuatIndex());
 membuatFolder("pages");
-if (!existsSync("pages/index.html") || !existsSync('pages/index.md')) {
+if (!existsSync("pages/index.html") && !existsSync('pages/index.md')) {
   writeFileSync("pages/index.html", `<h1>Hello World</h1>`);
 }
 // membuatFile("pages/index.html", /*html*/ `<h1>Hello World</h1>`);
